docs/cookbook/toy_codebase/pyproject.toml
[build-system]
requires = ["setuptools>=48", "setuptools_scm[toml]>=6.3.1"]
build-backend = "setuptools.build_meta"

[project]
name = "structured-qa"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
dynamic = ["version"]
dependencies = [
  "fire",
  "huggingface-hub",
  "llama-cpp-python",
  "loguru",
  "pydantic",
  "pymupdf4llm",
  "pyyaml",
  "rapidfuzz",
  "streamlit",
]

[project.optional-dependencies]
docs = [
  "mkdocs",
  "mkdocs-material",
  "mkdocstrings-python",
]

tests = [
  "pytest>=8,<9",
  "pytest-sugar>=0.9.6",
  "pytest-mock>=3.14.0"
]

[project.urls]
Documentation = "https://mozilla-ai.github.io/structured-qa/"
Issues = "https://github.com/mozilla-ai/structured-qa/issues"
Source = "https://github.com/mozilla-ai/structured-qa"

[tool.setuptools.packages.find]
exclude = ["tests", "tests.*"]
where = ["src"]
namespaces = false

[tool.setuptools_scm]

[project.scripts]
structured-qa = "structured_qa.cli:main"
